feat: add avatar module with CSV fallback
